requirements.txt
mysql-connector-python==8.0.33
pandas==1.3.5
scikit-learn==1.0.2
joblib==1.3.2
apache-airflow-providers-mysql==5.1.0
